pyproject.toml
[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
secondary = true

[tool.poetry]
name = "zillionare-omicron"
packages = [
    {include = "omicron"}
]
version = "2.0.0-alpha.18"
description = "Core Library for Zillionare"
authors = ["jieyu <code@jieyu.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://zillionare-omicron.readthedocs.io"
repository = "https://github.com/zillionare/omicron"
documentation = "https://zillionare-omicron.readthedocs.io"
keywords = ["AI", "quant", "trade", "stock"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.8",
]
include = [
    "LICENSE",
    "README.md",
    "HISTORY.md",
    "AUTHORS.md",
    "docs/*",
    "tests/*"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
sh = "1.14.1"
asyncpg = "0.21.0"
gino = "1.0.1"
aiohttp = ">=3.8.0"
cfg4py = ">=0.9"
arrow = "0.15.8"
numpy = "^1.22"
aioredis = "1.3.1"
aiosmtplib = "^1.1.6"
pandas = "^1.3.5"
zillionare-core-types = "^0.4.1"
ciso8601 = "^2.2.0"
Bottleneck = "^1.3.4"
Cython = "^0.29.28"
ZigZag = "^0.2.2"
scikit-learn = "^1.0.2"


black  = { version = "20.8b1", optional = true}
isort  = { version = "5.6.4", optional = true}
doc8  = { version = "0.8.1", optional = true}
flake8  = { version = "3.8.4", optional = true}
pre-commit  = { version = "2.8.2", optional = true}
pytest  = { version = "6.1.2", optional = true}
pytest-cov  = { version = "2.10.1", optional = true}
tox  = { version = "^3.20.1", optional = true}
virtualenv  = { version = "^20.2.2", optional = true}
pip  = { version = "^20.3.1", optional = true}
psutil  = { version = "^5.7.3", optional = true}
mkdocs  = { version = "^1.1.2", optional = true}
mkdocs-include-markdown-plugin  = { version = "^1.0.0", optional = true}
mkdocs-material  = { version = "^6.1.7", optional = true}
mkdocstrings  = { version = "^0.13.6", optional = true}
mkdocs-material-extensions  = { version = "^1.0.1", optional = true}
twine  = { version = "^3.3.0", optional = true}
idna = "2.5"
mkdocs-autorefs = {version = "^0.1.1", optional = true}
livereload = {version = "^2.6.3", optional = true}
empyrical-reloaded = "^0.5.8"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "doc8",
    "flake8",
    "pytest-cov",
    "psutil"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine"]

doc = [
    "mkdocs",
    "mkdocs-autorefs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "livereload"
    ]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
